backend_api/app/requirements.txt
fastapi
uvicorn
imaplib2
imapclient
google-auth
google_auth_oauthlib
msal
pydantic
pyjwt
passlib[bcrypt]
sqlalchemy
pymysql
databases
imapclient
bs4
httpx
google-api-python-client
fernet
python-dotenv